+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Updates
+
+- Update `rust-embed`: https://github.com/BBaoVanC/rust-embed/commit/227c4e0947c7d41d42af69685b66c3f93f578836
+
 ## [v0.2.0] - 2022-01-17
 
 ### Changed