+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Better keyboard support on JS-powered links (delete/cancel upload buttons)
   - rather than make the `keypress` event trigger the `click` event, just set an `href` and `click` magically works
+- Properly pluralize human durations (if 1 of the unit)
 
 ## [v0.2.1] - 2022-01-21